requirements.txt
requests
lxml
bs4
psutil
RPi.GPIO
Pillow
spidev